Replace error view-functions to core.views
